requirements.txt
atproto==0.0.44
jsprint==0.0.3
PyYAML==6.0.1
colorama==0.4.6
guilded-webhook==0.3.1